requirements.txt
python-telegram-bot==20.7
openai==0.28.0
python-dotenv==1.0.0